+++ scripts/macros/2024/classFeatures/monk/superiorDefense.js
@@ -1,0 +1,26 @@
+import {itemUtils} from '../../../../utils.js';
+async function added({trigger: {entity: item}}) {
+    await itemUtils.correctActivityItemConsumption(item, ['use'], 'monksFocus');
+}
+export let superiorDefense = {
+    name: 'Superior Defense',
+    version: '1.3.162',
+    rules: 'modern',
+    item: [
+        {
+            pass: 'created',
+            macro: added,
+            priority: 55
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 55
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 55
+        }
+    ]
+};